requirements.txt
transformers==4.49.0
torch
timm
einops
pillow
accelerate